requirements.txt
python==3.11.5
django==4.2.5
djangorestframework==3.14.0